Restructure flask and mock
