+++ openapi_core/contrib/flask/__init__.py
@@ -1,0 +1,4 @@
+from openapi_core.contrib.flask.requests import FlaskOpenAPIRequest
+from openapi_core.contrib.flask.responses import FlaskOpenAPIResponse
+
+__all__ = ['FlaskOpenAPIRequest', 'FlaskOpenAPIResponse']
