+++ tests/integration/validation/test_minimal.py
@@ -3,8 +3,8 @@ import pytest
 from openapi_core.schema.operations.exceptions import InvalidOperation
 from openapi_core.schema.paths.exceptions import InvalidPath
 from openapi_core.shortcuts import create_spec
+from openapi_core.testing.mock import MockRequest
 from openapi_core.validation.request.validators import RequestValidator
-from openapi_core.wrappers.mock import MockRequest
 
 
 class TestMinimal(object):
