+++ tests/integration/validation/test_petstore.py
@@ -16,10 +16,10 @@ from openapi_core.schema.schemas.enums import SchemaType
 from openapi_core.schema.schemas.exceptions import InvalidSchemaValue
 from openapi_core.schema.servers.exceptions import InvalidServer
 from openapi_core.shortcuts import create_spec
+from openapi_core.testing.mock import MockRequest, MockResponse
 from openapi_core.validation.request.datatypes import RequestParameters
 from openapi_core.validation.request.validators import RequestValidator
 from openapi_core.validation.response.validators import ResponseValidator
-from openapi_core.wrappers.mock import MockRequest, MockResponse
 
 
 class TestPetstore(object):
