+++ tests/integration/validation/test_validators.py
@@ -17,10 +17,10 @@ from openapi_core.schema.responses.exceptions import (
 )
 from openapi_core.schema.servers.exceptions import InvalidServer
 from openapi_core.shortcuts import create_spec
+from openapi_core.testing.mock import MockRequest, MockResponse
 from openapi_core.validation.request.datatypes import RequestParameters
 from openapi_core.validation.request.validators import RequestValidator
 from openapi_core.validation.response.validators import ResponseValidator
-from openapi_core.wrappers.mock import MockRequest, MockResponse
 
 
 class TestRequestValidator(object):
